[BpkLayout] Merge BpkArkProvider and BpkLayoutProvider into BpkProvider
Consolidate three provider exports into one. Consumers only need to
import and use BpkProvider — BpkArkProvider and BpkLayoutProvider are
no longer part of the public API.

- Inline BpkArkProvider logic (useArkLocale, LocaleProvider) into BpkProvider
- Delete BpkArkProvider.tsx and BpkLayoutProvider.tsx
- Remove BpkArkProvider and BpkLayoutProvider from index.ts exports
- Rename BpkArkProvider-test.tsx → BpkProvider-locale-test.tsx
- Delete BpkLayoutProvider-test.tsx (passthrough with no logic to test)
